Address CR Comments (#102)
