+++ app/src/main/cpp/skyline/services/am/controller/ISelfController.cpp
@@ -3,7 +3,6 @@
 
 #include <os.h>
 #include <services/hosbinder/GraphicBufferProducer.h>
-#include <services/hosbinder/display.h>
 #include "ISelfController.h"
 
 namespace skyline::service::am {
