--- app/src/main/cpp/skyline/services/hosbinder/buffer.h
@@ -1,12 +1,0 @@
-// SPDX-License-Identifier: MPL-2.0
-// Copyright © 2020 Skyline Team and Contributors (https://github.com/skyline-emu/)
-
-#pragma once
-
-#include <common.h>
-#include <services/common/parcel.h>
-#include <services/nvdrv/devices/nvmap.h>
-#include <gpu.h>
-
-namespace skyline::service::hosbinder {
-}
