+++ app/src/main/cpp/skyline/services/visrv/IDisplayService.cpp
@@ -3,7 +3,6 @@
 
 #include <os.h>
 #include <services/hosbinder/GraphicBufferProducer.h>
-#include <services/hosbinder/display.h>
 #include "IDisplayService.h"
 
 namespace skyline::service::visrv {
